Moved TGA decoder & encoder-specific code
