+++ src/tga/encoder.rs
@@ -1,9 +1,40 @@
-use std::io::Write;
+use std::{error, fmt, io::{self, Write}};
 
-use super::header::Header;
+use super::header::{Header, ImageType, ALPHA_BIT_MASK, SCREEN_ORIGIN_BIT_MASK};
 use crate::color::ColorType;
-use crate::error::ImageResult;
-use crate::image::ImageEncoder;
+use crate::error::{EncodingError, ImageResult, UnsupportedError, UnsupportedErrorKind};
+use crate::{image::ImageEncoder, ImageError, ImageFormat};
+use byteorder::{LittleEndian, WriteBytesExt};
+use std::convert::TryFrom;
+
+/// Errors that can occur during encoding and saving of a TGA image.
+#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
+enum EncoderError {
+    /// Invalid TGA width.
+    WidthInvalid(u32),
+
+    /// Invalid TGA height.
+    HeightInvalid(u32),
+}
+
+impl fmt::Display for EncoderError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            EncoderError::WidthInvalid(s) => f.write_fmt(format_args!("Invalid TGA width: {}", s)),
+            EncoderError::HeightInvalid(s) => {
+                f.write_fmt(format_args!("Invalid TGA height: {}", s))
+            }
+        }
+    }
+}
+
+impl From<EncoderError> for ImageError {
+    fn from(e: EncoderError) -> ImageError {
+        ImageError::Encoding(EncodingError::new(ImageFormat::Tga.into(), e))
+    }
+}
+
+impl error::Error for EncoderError {}
 
 /// TGA encoder.
 pub struct TgaEncoder<W: Write> {
@@ -61,10 +92,68 @@ impl<W: Write> ImageEncoder for TgaEncoder<W> {
     }
 }
 
+impl Header {
+    /// Load the header with values from pixel information.
+    pub(crate) fn from_pixel_info(
+        color_type: ColorType,
+        width: u32,
+        height: u32,
+    ) -> ImageResult<Self> {
+        let mut header = Self::default();
+
+        if width > 0 && height > 0 {
+            header.image_width = u16::try_from(width)
+                .map_err(|_| ImageError::from(EncoderError::WidthInvalid(width)))?;
+
+            header.image_height = u16::try_from(height)
+                .map_err(|_| ImageError::from(EncoderError::HeightInvalid(height)))?;
+
+            let (num_alpha_bits, other_channel_bits, image_type) = match color_type {
+                ColorType::Rgba8 | ColorType::Bgra8 => (8, 24, ImageType::RawTrueColor),
+                ColorType::Rgb8 | ColorType::Bgr8 => (0, 24, ImageType::RawTrueColor),
+                ColorType::La8 => (8, 8, ImageType::RawGrayScale),
+                ColorType::L8 => (0, 8, ImageType::RawGrayScale),
+                _ => {
+                    return Err(ImageError::Unsupported(
+                        UnsupportedError::from_format_and_kind(
+                            ImageFormat::Tga.into(),
+                            UnsupportedErrorKind::Color(color_type.into()),
+                        ),
+                    ))
+                }
+            };
+
+            header.image_type = image_type as u8;
+            header.pixel_depth = num_alpha_bits + other_channel_bits;
+            header.image_desc = num_alpha_bits & ALPHA_BIT_MASK;
+            header.image_desc |= SCREEN_ORIGIN_BIT_MASK; // Upper left origin.
+        }
+
+        Ok(header)
+    }
+
+    /// Write out the header values.
+    pub(crate) fn write_to(&self, w: &mut dyn Write) -> io::Result<()> {
+        w.write_u8(self.id_length)?;
+        w.write_u8(self.map_type)?;
+        w.write_u8(self.image_type)?;
+        w.write_u16::<LittleEndian>(self.map_origin)?;
+        w.write_u16::<LittleEndian>(self.map_length)?;
+        w.write_u8(self.map_entry_size)?;
+        w.write_u16::<LittleEndian>(self.x_origin)?;
+        w.write_u16::<LittleEndian>(self.y_origin)?;
+        w.write_u16::<LittleEndian>(self.image_width)?;
+        w.write_u16::<LittleEndian>(self.image_height)?;
+        w.write_u8(self.pixel_depth)?;
+        w.write_u8(self.image_desc)?;
+        Ok(())
+    }
+}
+
 #[cfg(test)]
 mod tests {
-    use super::TgaEncoder;
-    use crate::tga::{header::EncoderError, TgaDecoder};
+    use super::{EncoderError, TgaEncoder};
+    use crate::tga::TgaDecoder;
 
     use crate::color::ColorType;
     use crate::image::ImageDecoder;
