+++ src/tga/header.rs
@@ -1,45 +1,5 @@
-use crate::{
-    error::{EncodingError, UnsupportedError, UnsupportedErrorKind},
-    ColorType, ImageError, ImageFormat, ImageResult,
-};
-use byteorder::{LittleEndian, ReadBytesExt, WriteBytesExt};
-use std::convert::TryFrom;
-use std::{
-    error, fmt,
-    io::{Read, Write},
-};
-
 pub(crate) const ALPHA_BIT_MASK: u8 = 0b1111;
 pub(crate) const SCREEN_ORIGIN_BIT_MASK: u8 = 0b10_0000;
-
-/// Errors that can occur during encoding and saving of a TGA image.
-#[derive(Debug, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub(crate) enum EncoderError {
-    /// Invalid TGA width.
-    WidthInvalid(u32),
-
-    /// Invalid TGA height.
-    HeightInvalid(u32),
-}
-
-impl fmt::Display for EncoderError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            EncoderError::WidthInvalid(s) => f.write_fmt(format_args!("Invalid TGA width: {}", s)),
-            EncoderError::HeightInvalid(s) => {
-                f.write_fmt(format_args!("Invalid TGA height: {}", s))
-            }
-        }
-    }
-}
-
-impl From<EncoderError> for ImageError {
-    fn from(e: EncoderError) -> ImageError {
-        ImageError::Encoding(EncodingError::new(ImageFormat::Tga.into(), e))
-    }
-}
-
-impl error::Error for EncoderError {}
 
 pub(crate) enum ImageType {
     NoImageData = 0,
@@ -115,80 +75,4 @@ pub(crate) struct Header {
     pub(crate) image_height: u16,  // height of image
     pub(crate) pixel_depth: u8,    // bits per pixel
     pub(crate) image_desc: u8,     // image descriptor
-}
-
-impl Header {
-    /// Load the header with values from pixel information.
-    pub(crate) fn from_pixel_info(
-        color_type: ColorType,
-        width: u32,
-        height: u32,
-    ) -> ImageResult<Self> {
-        let mut header = Self::default();
-
-        if width > 0 && height > 0 {
-            header.image_width = u16::try_from(width)
-                .map_err(|_| ImageError::from(EncoderError::WidthInvalid(width)))?;
-
-            header.image_height = u16::try_from(height)
-                .map_err(|_| ImageError::from(EncoderError::HeightInvalid(height)))?;
-
-            let (num_alpha_bits, other_channel_bits, image_type) = match color_type {
-                ColorType::Rgba8 | ColorType::Bgra8 => (8, 24, ImageType::RawTrueColor),
-                ColorType::Rgb8 | ColorType::Bgr8 => (0, 24, ImageType::RawTrueColor),
-                ColorType::La8 => (8, 8, ImageType::RawGrayScale),
-                ColorType::L8 => (0, 8, ImageType::RawGrayScale),
-                _ => {
-                    return Err(ImageError::Unsupported(
-                        UnsupportedError::from_format_and_kind(
-                            ImageFormat::Tga.into(),
-                            UnsupportedErrorKind::Color(color_type.into()),
-                        ),
-                    ))
-                }
-            };
-
-            header.image_type = image_type as u8;
-            header.pixel_depth = num_alpha_bits + other_channel_bits;
-            header.image_desc = num_alpha_bits & ALPHA_BIT_MASK;
-            header.image_desc |= SCREEN_ORIGIN_BIT_MASK; // Upper left origin.
-        }
-
-        Ok(header)
-    }
-
-    /// Load the header with values from the reader.
-    pub(crate) fn from_reader(r: &mut dyn Read) -> ImageResult<Self> {
-        Ok(Self {
-            id_length: r.read_u8()?,
-            map_type: r.read_u8()?,
-            image_type: r.read_u8()?,
-            map_origin: r.read_u16::<LittleEndian>()?,
-            map_length: r.read_u16::<LittleEndian>()?,
-            map_entry_size: r.read_u8()?,
-            x_origin: r.read_u16::<LittleEndian>()?,
-            y_origin: r.read_u16::<LittleEndian>()?,
-            image_width: r.read_u16::<LittleEndian>()?,
-            image_height: r.read_u16::<LittleEndian>()?,
-            pixel_depth: r.read_u8()?,
-            image_desc: r.read_u8()?,
-        })
-    }
-
-    /// Write out the header values.
-    pub(crate) fn write_to(&self, w: &mut dyn Write) -> ImageResult<()> {
-        w.write_u8(self.id_length)?;
-        w.write_u8(self.map_type)?;
-        w.write_u8(self.image_type)?;
-        w.write_u16::<LittleEndian>(self.map_origin)?;
-        w.write_u16::<LittleEndian>(self.map_length)?;
-        w.write_u8(self.map_entry_size)?;
-        w.write_u16::<LittleEndian>(self.x_origin)?;
-        w.write_u16::<LittleEndian>(self.y_origin)?;
-        w.write_u16::<LittleEndian>(self.image_width)?;
-        w.write_u16::<LittleEndian>(self.image_height)?;
-        w.write_u8(self.pixel_depth)?;
-        w.write_u8(self.image_desc)?;
-        Ok(())
-    }
 }
